+++ src/screens/driver.js
@@ -2,7 +2,7 @@ import React, { useState, useContext, useEffect } from 'react';
 import { Alert } from 'react-native';
 import { withNavigation } from 'react-navigation';
 
-import { CContent, CBarCodeReader } from '@components';
+import { CContent, CBarCodeReader, CSpace, DEFAULT_FONT_SIZE, CText } from '@components';
 import { DriverContext } from '@contexts';
 import { NAVS } from '@screens';
 import Dialog from 'react-native-dialog';
@@ -10,6 +10,7 @@ import Dialog from 'react-native-dialog';
 /**
  * Should display a barcode reader to get driver infos
  * Display an Dialog Input to validate driver GMS
+ * For tests use BT00249316
  */
 const ScreenDriver = ({ navigation }) => {
   // manage the driver context
@@ -17,22 +18,23 @@ const ScreenDriver = ({ navigation }) => {
   // manage dialog gsm
   const [showGSMInput, setShowGSMInput] = useState(false);
   // manage gsm number entered
-  const [tempGsmNumber, setTempGsmNumber] = useState(driverContext.gsmNumber);
+  const [tempGsmNumber, setTempGsmNumber] = useState(driverContext.driver.gsm);
   // manage hiding the barcode
   const [hideBarCodeReader, setHideBarCodeReader] = useState(false);
 
   // manage modals show/hide
   useEffect(() => {
-    if (tempGsmNumber.length === 0 && driverContext.gsmNumber.length > 0) {
-      setTempGsmNumber(driverContext.gsmNumber);
+    if (tempGsmNumber.length === 0 && driverContext.driver.gsm.length > 0) {
+      setTempGsmNumber(driverContext.driver.gsm);
     }
     setHideBarCodeReader(showGSMInput);
-  });
+  }, [tempGsmNumber, driverContext]);
 
   // if successfully read a bar code then display GSM modal
   // need to timed for display... probably due to setState
-  const onBarCodeSuccess = () => {
+  const onBarCodeSuccess = value => {
     setTimeout(() => {
+      setHideBarCodeReader(true);
       setShowGSMInput(true);
     }, 500);
   };
@@ -40,7 +42,10 @@ const ScreenDriver = ({ navigation }) => {
   // if error on read show it
   const onBarCodeError = value => {
     setTimeout(() => {
-      Alert.alert(value);
+      if (value.error) Alert.alert(value.error);
+      else Alert.alert(value);
+      setHideBarCodeReader(false);
+      setShowGSMInput(false);
     }, 500);
   };
 
@@ -54,9 +59,7 @@ const ScreenDriver = ({ navigation }) => {
       .then(() => {
         setShowGSMInput(false);
         setHideBarCodeReader(true);
-        setTimeout(() => {
-          navigation.navigate(NAVS.driver.next);
-        }, 500);
+        navigation.navigate(NAVS.driver.next);
       })
       .catch(() => {
         Alert.alert('Impossible de sauvegarder votre numéro...');
@@ -65,6 +68,7 @@ const ScreenDriver = ({ navigation }) => {
 
   const onPressCancelInputGSM = () => {
     setShowGSMInput(false);
+    setHideBarCodeReader(false);
     setTempGsmNumber('');
   };
 
@@ -75,9 +79,13 @@ const ScreenDriver = ({ navigation }) => {
 
   return (
     <DriverContext.Consumer>
-      {({ gsmNumber, firstname, lastname, getDriverDatas }) => (
+      {({ driver, getDriverDatas }) => (
         <>
           <CContent title="Code de tournée" fullscreen pressBackHome={onPressBackHome}>
+            <CText style={{ textAlign: 'center', fontSize: (DEFAULT_FONT_SIZE * 3) / 4 }}>
+              {`Scannez le code barre de votre tournée.`}
+            </CText>
+            <CSpace />
             <CBarCodeReader
               testID="ID_BARCODE"
               verificator={getDriverDatas}
@@ -87,13 +95,15 @@ const ScreenDriver = ({ navigation }) => {
             />
           </CContent>
           <Dialog.Container visible={showGSMInput} testID="ID_GSMCONFIRM">
-            <Dialog.Title>{`${firstname} ${lastname}`}</Dialog.Title>
+            <Dialog.Title>{`${driver.firstname} ${driver.lastname}`}</Dialog.Title>
             <Dialog.Description>
-              {`Veuillez confirmer votre numéro de portable ({tempGsmNumber}) afin d'être joint par votre superviseur.`}
+              {`Veuillez confirmer votre numéro de portable (${
+                driver.gsm
+              }) afin d'être joint par votre superviseur.`}
             </Dialog.Description>
             <Dialog.Input
               testID="ID_GSMCONFIRM_INPUT"
-              defaultValue={gsmNumber}
+              defaultValue={driver.gsm}
               numberOfLines={1}
               onChangeText={onChangeInputGSM}
             />